+++ packages/devtools_app/lib/src/screens/network/network_screen.dart
@@ -205,7 +205,6 @@ class _NetworkProfilerControlsState extends State<_NetworkProfilerControls>
     }
 
     final screenWidth = ScreenSize(context).width;
-    final hasRequests = controller.filteredData.value.isNotEmpty;
     return Column(
       children: [
         Row(
@@ -233,7 +232,7 @@ class _NetworkProfilerControlsState extends State<_NetworkProfilerControls>
             Expanded(
               child: SearchField<NetworkController>(
                 searchController: controller,
-                searchFieldEnabled: hasRequests,
+                searchFieldEnabled: _recording,
                 searchFieldWidth: screenWidth <= MediaSize.xs
                     ? defaultSearchFieldWidth
                     : wideSearchFieldWidth,